added fake store API connection

+++ src/services/server/fakestore.js
@@ -27,7 +27,7 @@ function fakeStoreToKlarnaCart(fakeStoreCart) {
 			name: fakeStoreItem.title,
 			quantity: fakeStoreItem.quantity,
 			quantity_unit: 'pcs', // same
-			unit_price: parseFloat(fakeStoreItem.price) * 100,
+			unit_price: parseInt(fakeStoreItem.price) * 100,
 			tax_rate: 1000, // same
 			total_discount_amount: 0 // same
 		};
